real_estate_app/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-Limiter==3.12
psycopg2-binary==2.9.10
python-dotenv==1.0.1
werkzeug==3.1.1
gunicorn==22.0.0